Fix: remove dtype param
Former-commit-id: 3bcf8ec37237ef3ebbf5dc8517f1ef33ed018adb

+++ ivy/functional/backends/numpy/statistical.py
@@ -34,9 +34,7 @@ def mean(
     out: Optional[np.ndarray] = None,
 ) -> np.ndarray:
     axis = tuple(axis) if isinstance(axis, list) else axis
-    return np.asarray(
-        np.mean(x, axis=axis, dtype="float32", keepdims=keepdims, out=out)
-    )
+    return np.asarray(np.mean(x, axis=axis, keepdims=keepdims, out=out))
 
 
 mean.support_native_out = True
